Move setup of message bus to ApplicationContainerCluster
