+++ config-model/src/main/java/com/yahoo/vespa/model/admin/clustercontroller/ClusterControllerContainerCluster.java
@@ -27,4 +27,6 @@ public class ClusterControllerContainerCluster extends ContainerCluster<ClusterC
     @Override
     protected void doPrepare(DeployState deployState) { }
 
+    protected boolean messageBusEnabled() { return false; }
+
 }
